[AAE-1137] Refactor delete process directive
