+++ demo-shell/src/app/components/cloud/process-details-cloud-demo.component.html
@@ -1,7 +1,15 @@
-
+<div class="app-process-cloud-buttons-container">
 <button data-automation-id="go-back" mat-icon-button (click)="onGoBack()">
     <mat-icon>arrow_back</mat-icon> Go Back
 </button>
+
+<button mat-icon-button adf-cloud-cancel-process data-automation-id="cancel-process"
+        (success)="onCancelSuccess()" (error)="onCancelError($event)">
+    <mat-icon>delete</mat-icon>
+    Cancel process
+</button>
+<mat-error *ngIf="errorMessage">{{errorMessage}}</mat-error>
+</div>
 
 <h4 data-automation-id="process-details-header">Simple page to show the process instance: {{ processInstanceId }} of the app: {{ appName }}</h4>
 
@@ -22,4 +30,3 @@
         [processInstanceId]="processInstanceId">
     </adf-cloud-process-header>
 </div>
-
