+++ demo-shell/src/app/components/cloud/process-details-cloud-demo.component.ts
@@ -27,6 +27,8 @@ export class ProcessDetailsCloudDemoComponent {
     processInstanceId: string;
     appName: string;
 
+    errorMessage: string;
+
     constructor(private route: ActivatedRoute, private router: Router) {
         this.route.params.subscribe((params) => {
             this.processInstanceId = params.processInstanceId;
@@ -45,5 +47,14 @@ export class ProcessDetailsCloudDemoComponent {
         if (taskId) {
             this.router.navigate([`/cloud/${this.appName}/task-details/${taskId}`]);
         }
+    }
+
+    onCancelSuccess() {
+        this.errorMessage = '';
+        this.onGoBack();
+    }
+
+    onCancelError($event) {
+        this.errorMessage = $event;
     }
 }
